Fixed CUDA bugs

+++ Cuda_NB/classifier.cuh
@@ -21,18 +21,18 @@ class GaussianNB {
 public:
   unsigned int n_features_ = 1; // Number of features
   unsigned int n_classes_ = 1;  // Number of unique labels
-  double *feature_means_;       // Per feature per class means
-  double *feature_vars_;        // Per feature per class variances
-  double *class_priors_;        // shape: n_classes_
+  float *feature_means_;       // Per feature per class means
+  float *feature_vars_;        // Per feature per class variances
+  float *class_priors_;        // shape: n_classes_
   int *class_count_;            // Number of items per class; shape: n_classes_
 
   GaussianNB();
 
   virtual ~GaussianNB();
 
-  void train(vector<double> data, vector<int> labels);
+  void train(vector<float> data, vector<int> labels);
 
-  int predict(vector<double> data, vector<int> labels);
+  int predict(vector<float> data, vector<int> labels);
 };
 
 class MultinomialNB {
@@ -41,33 +41,33 @@ public:
   vector<int>::size_type n_features_ = 1; // Number of features
   unsigned int n_classes_ = 1;            // Number of unique labels
 
-  double *feature_probs; // shape: n_classes_ * n_features_
-  double *class_priors;  // shape: n_classes_
+  float *feature_probs; // shape: n_classes_ * n_features_
+  float *class_priors;  // shape: n_classes_
 
   MultinomialNB();
 
   virtual ~MultinomialNB();
 
-  void train(vector<double> data, vector<int> labels);
+  void train(vector<float> data, vector<int> labels);
 
-  int predict(vector<double> data, vector<int> labels);
+  int predict(vector<float> data, vector<int> labels);
 };
 
 class BernoulliNB {
 
 public:
-  double *feature_probs;
+  float *feature_probs;
   vector<int>::size_type n_features_ = 1; // Number of features
   unsigned int n_classes_ = 1;            // Number of samples in each class
-  double *class_count_;                   // Class prior probabilities
+  float *class_count_;                   // Class prior probabilities
 
   BernoulliNB();
 
   virtual ~BernoulliNB();
 
-  void train(vector<double> data, vector<int> labels);
+  void train(vector<float> data, vector<int> labels);
 
-  int predict(vector<double> data, vector<int> labels);
+  int predict(vector<float> data, vector<int> labels);
 };
 
 class ComplementNB {
@@ -75,19 +75,19 @@ class ComplementNB {
 public:
   vector<int>::size_type n_features_ = 1; // Number of features
   unsigned int n_classes_ = 1;            // Number of samples in each class
-  double *feature_weights_; // Learned weights of the model; n_features_ * n_classes_
-  double *per_class_feature_sum_; // n_features_ * n_classes_
-  double *per_feature_sum_;
-  double *per_class_sum_;
-  double all_sum_;
+  float *feature_weights_; // Learned weights of the model; n_features_ * n_classes_
+  float *per_class_feature_sum_; // n_features_ * n_classes_
+  float *per_feature_sum_;
+  float *per_class_sum_;
+  float all_sum_;
 
   ComplementNB();
 
   virtual ~ComplementNB();
 
-  void train(vector<double> data, vector<int> labels);
+  void train(vector<float> data, vector<int> labels);
 
-  int predict(vector<double> data, vector<int> labels);
+  int predict(vector<float> data, vector<int> labels);
 };
 
 #endif
